% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbnorm.R
\name{ACDdbnorm}
\alias{ACDdbnorm}
\title{Adjusted coefficient of determination for a data normalized for across batch signal drift}
\usage{
ACDdbnorm(m)
}
\arguments{
\item{m}{A data frame in which rows define the independent experiments (samples) and columns the features (variables), with the batch levels in the first column.}
}
\value{
Several graphs compiled into a \bold{PDF} file which are a \emph{correlation} plot for each of applied models, a grouped \emph{barplot} presenting the maximum variability associated with batch levels in the raw and the corrected datasets.\cr
Files saved as \bold{csv} in the working directory are a dataset corrected via either of applied models. Also, a two column matrix for Adjusted R-Square for raw and corrected datasets and a table summarizing the score values presented in \emph{barplot}.
}
\description{
This function gives a quick notification about the performance of the statistical models implemented in the \code{dbnorm} package such as \emph{Giordan (2013)} and/or empirical Bayes methods in two setting of parametric and non-parametric as described by \emph{Johnson et al.(2007)} and in \emph{sva} package by \emph{Leek et al.(2012)}. It calculates and plots adjusted coefficient of determination or \emph{Adjusted R-Squared} for each variable estimated in a regression model for its dependency to the batch level in the raw data and treated data via either of those models. Immediately, a score calculated based on the maximum variability explained by the batch level presents the performance of applied models. This score notifies the consistency of a model performance for all detected features (variables), facilitating quick comparison of the models for selecting one of those models, which is more appropriate to the data structure. This function is suggested for less than 2000 features (variables) to keep maximum computational speed.\cr
}
\details{
Zero and NA values are not allowed. Optionally missing value can be imputed by \code{emvf} or \code{emvd}, functions implemented in \code{dbnorm} package. Input data must be normalized prior.
}
\examples{

batch<- rep(gl(3,7,labels = c(1:3)),1)
y<- matrix(rnorm(2100),nrow=21)
m<-data.frame(batch,y)
}
\references{
M.Giordan (2013) \emph{https://link.springer.com/article/10.1007/s12561-013-9081-1}\cr
Johnson et al. (2007) \emph{http://www.ncbi.nlm.nih.gov/pubmed/16632515}\cr
Leek et al. (2012) \emph{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3307112/}
}
\keyword{Adjusted}
\keyword{R-square}
