% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNorm_Mapping.R
\name{conNormalizer}
\alias{conNormalizer}
\title{Normalizing a target data array to a basis array based on their distributions}
\usage{
conNormalizer(tg, bs)
}
\arguments{
\item{tg}{a target data array}

\item{bs}{a basis data array}
}
\value{
A normalized target data array with 
the same distribution with the basis data array
}
\description{
Normalizing a target data array to a basis array based on their distributions
}
\details{
The function maps a target data array to a basis array based on their distributions 
and the basis data array can be an arbitary data array or a standard distribution 
such as normal distribution.
}
\examples{
# Normalize DArray1 to DArray3
# load build-in data arrays 
data(DArray1) 
data(DArray3) 

# Capturing distribution information 
DBdata1 <- genDistData(DArray1, 500)
DBdata3 <- genDistData(DArray3, 500)

# Using Gaussian function to fit DBdata3
DBdata3 <- gaussianFit(DBdata3)

# Normalize DBdata1 to the Gaussian fitting function of DBdata3
DArray1 = conNormalizer(DArray1, DArray3)
DA1toDA3DBdata <- genDistData(DA1toDA3, 500)
visDistData(DA1toDA3DBdata, "P", "DA1toDA3", "Range", "Probability")

}
\author{
Qinxue Meng, Paul Kennedy
}

