% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNorm_Fitting.R
\name{custFit}
\alias{custFit}
\title{fitting a distribution by a customised curve function}
\usage{
custFit(DBdata, formula)
}
\arguments{
\item{DBdata}{input distribution dataset}

\item{formula}{a customised curve function}
}
\value{
a customised curve fitting function
}
\description{
fitting a distribution by a customised curve function
}
\details{
The function fits distributions by a customised curve fitting and
returns a customised curve fitting function.
}
\examples{
# Calculating the customised curve fitting function of DArray1's distribution
DBdata1 = custFit(DBdata1)

}
\author{
Qinxue Meng, Paul Kennedy
}
\seealso{
\code{lm}
}

