\name{DBNorm_package-package}
\alias{DBNorm_package-package}
\alias{DBNorm_package}
\docType{package}
\title{
\packageTitle{DBNorm_package}
}
\description{
\packageDescription{DBNorm_package}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DBNorm_package}
\packageIndices{DBNorm_package}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{DBNorm_package}

Maintainer: \packageMaintainer{DBNorm_package}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
