% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbnorm.R
\name{dbnormPcom}
\alias{dbnormPcom}
\title{Drift Across Batch Normalization via Parametric- ComBat model and visualization}
\usage{
dbnormPcom(m)
}
\arguments{
\item{m}{A data frame in which rows define the independent experiments (samples) and columns the features (variables), with the batch levels in the first column.}
}
\value{
Several graphs compiled into a \bold{PDF} file are a \emph{PCA} score plot, \emph{Scree} plot and a  \emph{correlation} plot for raw and corrected data. Also, the \emph{RLA} plots for each dataset visualized in the \bold{Viewer} panel in the \bold{rstudio} console.\cr
Files saved as \bold{csv} in the working directory are a dataset corrected by the applied model. Also, a two column matrix for Adjusted R-Square raw and corrected dataset and a table summarizing the maximum score.
}
\description{
This function allows you adjust the data for signal drift across multiple batches or batch effect via parametric Empirical Bayes approach as described by Johnson et al. 2007 ( see also “sva”). \emph{dbnormPcom} includes advanced statistical tools to inspect the structure and quality of high throughput experiment both in macroscopic and microscopic scales at the level of sample sets and metabolic feature, respectively. Notably, using this function users applied unsupervised learning algorithm to visualize the most variance explained by the two first components in the different set of samples analyzed in the entire experiment in the raw and corrected data. In parallel, linear association of feature (variable) and batch level has been estimated and visualized by a correlation plot. In fact, estimated \emph{Adjusted- R squared} is considered to define the level of dependency of feature (variable) to the batch level in the raw and corrected datasets. Besides, for quick notification about the performance of the applied model a maximum variability detected in either of datasets is reported as a score. This score notify the consistency of model performance for all detected features (variables).\cr
}
\details{
Zero and NA values are not allowed. Optionally, missing value can be imputed by the functions such as \code{emvf} and /or \code{emvd} implemented in \code{dbnorm} package. Input must be normalized and transformed prior.
}
\examples{
\dontrun{
batch<- rep(gl(5,10,labels = c(1:5)),1)
y<- matrix(rnorm(5000),nrow=50)
m<-data.frame(batch,y)
dbnormPcom(m)}
}
\references{
Johnson et al. 2007 \url{http://www.ncbi.nlm.nih.gov/pubmed/16632515}\cr
Leek et al., (2012) \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3307112/}
}
\keyword{ComBat-Parametric}
\keyword{analysis}
\keyword{and}
\keyword{data}
\keyword{for}
\keyword{normalized}
\keyword{regression}
\keyword{unsupervised}
\keyword{via}
