% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNorm_commons.R
\name{defineDist}
\alias{defineDist}
\title{Generating distribution data based on predefined distribution}
\usage{
defineDist(dist, min = 0, max = 1)
}
\arguments{
\item{dist}{a predefined distribution}

\item{min}{the lower bound of data range and default value is 0}

\item{max}{the upper bound of data range and default value is 1}
}
\value{
a distribution dataset of the input predefined distribution
}
\description{
Generating distribution data based on predefined distribution
}
\details{
This function generates distribution data based on predefined distribution. 
The purpose of this function is to enable to normalize arbitrary distributions 
into a standard distribution.
}
\examples{
# generate distribution data of a normal distribution
DArray5 <- defineDist(Norm(mean=0, sd=1))

}
\author{
Qinxue Meng, Paul Kennedy
}
\seealso{
\code{list()}
}

