% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNorm_Mapping.R
\name{disNormalizer}
\alias{disNormalizer}
\title{Normalizing a target data array to a basis array based on element positions}
\usage{
disNormalizer(tg, bs)
}
\arguments{
\item{tg}{a target data array}

\item{bs}{a basis data array}
}
\value{
A normalized target data array with 
the same distribution with the basis data array
}
\description{
Normalizing a target data array to a basis array based on element positions
}
\details{
The function normalize target data array to a basis array based on element 
positions. This method does not need to do fitting before normalization and 
works for discrete values as well.
}
\examples{
# Calculating the polynomial curve fitting function of DArray1's distribution
DArray1 = disNormalizer(DArray1, DArray3)

}
\author{
Qinxue Meng, Paul Kennedy
}

