% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNorm_Mapping.R
\name{distrNormalizer}
\alias{distrNormalizer}
\title{Normalizing a target data array to a standard distribution}
\usage{
distrNormalizer(tg, bs)
}
\arguments{
\item{tg}{a target data array}

\item{bs}{a standard distribution created by defineDist(dist)}
}
\value{
A normalized target data array with 
the same distribution with the standard distribution
}
\description{
Normalizing a target data array to a standard distribution
}
\details{
The function normalize target data array to a standard distribution.
}
\examples{
# Normalize a given data array into a normal distribution
loadData(0)
DBdata1 <- genDistData(DArray1, 500)
DBdata5 <- defineDist(Norm(mean=0, sd=1))
DA1toDA5 <- distrNormalizer(DBdata1, DBdata5)
DA1toDA5DBdata <- genDistData(DA1toDA5, 500)
visDistData(DA1toDA5DBdata, "P", "DA1toDA5", "Range", "Probability")

}
\author{
Qinxue Meng, Paul Kennedy
}

