% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbnorm.R
\name{emvd}
\alias{emvd}
\title{Estimation of missing value data-based}
\usage{
emvd(m)
}
\arguments{
\item{m}{An array or a matrix}
}
\value{
A matrix with estimated missing value.
}
\description{
This is a function in the \code{dbnorm}, a package in R. It returns to a matrix of data in which missing values are estimated by the lowest detected value in the entire experiment. By this function, all NA values are replaced by Zero values, that of being ultimately replaced by the lowest value detected in the experiment. Ultimately, data matrix is transposed to restore original structure.\cr
}
\details{
empty entries are not allowed
}
\examples{
m<- data.frame(x1=c(50,NA,6,10,30),x2=c(2,8,NA,15,0))
emvd(m)

}
\keyword{Missing}
\keyword{estimation}
\keyword{value}
