% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbnorm.R
\name{emvf}
\alias{emvf}
\title{Estimation of missing value feature-based}
\usage{
emvf(m)
}
\arguments{
\item{m}{An array or a matrix}
}
\value{
A matrix with estimated missing value.
}
\description{
It returns to a matrix of data in which missing values (Zero and/or NA values) are estimated. By this function, all Zero values are first replaced by NA values, which are then replaced by the lowest detected value on the column margin.
}
\details{
empty entries are not allowed
}
\examples{
m<- data.frame(x1=c(50,NA,6,10,30),x2=c(2,8,NA,15,0))

}
\keyword{Missing}
\keyword{estimation}
\keyword{value}
