% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNorm_Fitting.R
\name{fourierFit}
\alias{fourierFit}
\title{fitting a distribution by fourier curve fitting}
\usage{
fourierFit(DBdata, n)
}
\arguments{
\item{DBdata}{input distribution dataset}

\item{n}{the degree of the fourier fitting function}
}
\value{
a fourier curve fitting function
}
\description{
fitting a distribution by fourier curve fitting
}
\details{
The function fits distributions by fourier curve fitting and
returns a fourier curve fitting function.
}
\examples{
# Calculating the fourier curve fitting function of DArray1's distribution
DBdata1 = fourierFit(DBdata1, 3)

}
\author{
Qinxue Meng, Paul Kennedy
}
\seealso{
\code{lm}
}

