% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNorm_commons.R
\name{genDistData}
\alias{genDistData}
\title{Generating distribution dataset based on input data arrays.}
\usage{
genDistData(data, nbin)
}
\arguments{
\item{data}{input data array}

\item{nbin}{number of bins}
}
\value{
a distribution dataset of a given input data array
}
\description{
Generating distribution dataset based on input data arrays.
}
\details{
This function generates distribution dataset based on input data arrays for downstream analysis.
}
\examples{
# load DArray1
DData1 <- genDistData(DArray1, 500)
}
\author{
Qinxue Meng, Paul Kennedy
}
\seealso{
\code{list()}
}

