% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNorm_LoadData.R
\name{loadData}
\alias{loadData}
\title{Loading build-in datasets}
\usage{
loadData(n)
}
\arguments{
\item{n}{n-th data array to load; 
if n = 1, DArray1 is loaded; 
if n = 2, DArray2 is loaded; 
if n = 3, DArray3 is loaded; 
if n = 4, DArray4 is loaded; 
if n is not 1, 2, 3, 4, all four data arrays are loaded.}
}
\value{
None
}
\description{
This function loads build-in data array for examples
}
\details{
This function loads example data arrays for user to test
}
\examples{
# load DArray1
loadData(1)
# load all data arrays
loadData(5)
}
\author{
Qinxue Meng
}
\seealso{
\code{data()}
}

