% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbnorm.R
\name{profplotber}
\alias{profplotber}
\title{Visualization of analytical heterogeneity on the profile of features (variables) in ber- corrected data}
\usage{
profplotber(m)
}
\arguments{
\item{m}{A data frame in which rows define the independent experiments (samples) and columns the features (variables), with the batch in the first column.}
}
\value{
Original and adjusted datasets in \bold{csv} format together with the series of profile plot for the variables( features) in the sample sets analyzed in the entire experiment provided by the \emph{Scatter} plot,\emph{Violin} plot and \emph{pdf} plot compiled into \bold{PDF} file.
}
\description{
\emph{profplotber} allows you to adjust the data for batch effect using two-stage procedure approach as describes by \emph{Giordan (2013)} and informs you about the presence of across batch signal drift or batch effect in the treated data determined by the shifted probability density function plots (\emph{pdf} plots) of features (variables) detected in an experiment.\cr
}
\details{
Zero and NA values are not allowed. Optionally missing value can be imputed by functions implemented in \code{dbnorm} package. Input must be normalized and transformed prior.
}
\examples{

batch<- rep(gl(3,7,labels = c(1:3),1))
y<- matrix(rnorm(2100),nrow=21)
m<-data.frame(batch,y)
}
\references{
M.Giordan (2013) \emph{https://link.springer.com/article/10.1007/s12561-013-9081-1}
}
\keyword{and}
\keyword{ber}
\keyword{correction}
\keyword{plot}
\keyword{profile}
