% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbnorm.R
\name{profplotnpcom}
\alias{profplotnpcom}
\title{Visualization of analytical heterogeneity on the profile of features (variables) in Non-Parametric ComBat corrected data}
\usage{
profplotnpcom(m)
}
\arguments{
\item{m}{A data frame in which rows define the independent experiments (samples) and columns the features (variables), with the batch in the first column.}
}
\value{
Original and adjusted datasets in \bold{csv} format together with the series of profile plot of the features (variables) in the sample sets provided by \emph{Scatter} plot,\emph{Violin} plot and \emph{pdf} plot compiled into a \bold{PDF} file.
}
\description{
Visualization of analytical heterogeneity on the profile of features (variables) in Non-Parametric ComBat corrected data
}
\details{
Zero and NA values are not allowed. Optionally missing value can be imputed by functions implemented in \code{dbnorm} package. Input must be normalized and transformed prior.
}
\examples{

batch<- rep(gl(3,7,labels = c(1:3),1))
y<- matrix(rnorm(2100),nrow=21)
m<-data.frame(batch,y)

}
\references{
Johnson et al. (2007) \emph{http://www.ncbi.nlm.nih.gov/pubmed/16632515}\cr
Leek et al. (2012) \emph{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3307112/}
}
\keyword{ComBat}
\keyword{Non}
\keyword{Parametric}
\keyword{and}
\keyword{plot}
\keyword{profile}
