% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbnorm.R
\name{profplotpcom}
\alias{profplotpcom}
\title{Visualization of analytical heterogeneity on the profile of features (variables) in ComBat-Parametric -corrected data}
\usage{
profplotpcom(m)
}
\arguments{
\item{m}{A data frame in which rows define the independent experiments (samples) and columns the features (variables), with the batch in the first column.}
}
\value{
Original and adjusted datasets in \bold{csv} format together with the series of profile plot for the features(variables) in the sample sets provided by the \emph{Scatter} plot,\emph{Violin} plot and \emph{pdf} plot compiled into a \bold{PDF} file.
}
\description{
\emph{profplotpcom} allows you to adjust the data for batch effect using Parametric Empirical Bayes approach as described by \emph{Johnson et al.(2007)} and via \emph{sva} package as explained by \emph{Leek et al.(2012)}, and informs you about the presence of across batch signal drift or batch effect in the treated data, determined by the shifted probability density function plots (\emph{pdf} plots) of features (variables) detected in an experiment.\cr
}
\details{
Zero and NA values are not allowed. Optionally missing value can be imputed functions implemented in \code{dbnorm} package. Input must be normalized and transformed prior.
}
\examples{

batch<- rep(gl(3,7,labels = c(1:3),1))
y<- matrix(rnorm(2100),nrow=21)
m<-data.frame(batch,y)
}
\references{
Johnson et al. (2007) \emph{http://www.ncbi.nlm.nih.gov/pubmed/16632515}\cr
Leek et al. (2012) \emph{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3307112/}
}
\keyword{ComBat}
\keyword{Parametric}
\keyword{and}
\keyword{plot}
\keyword{profile}
