% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbnorm.R
\name{Visdbnorm}
\alias{Visdbnorm}
\title{Visualization and normalization of signal drift across batches\cr}
\usage{
Visdbnorm(f)
}
\arguments{
\item{f}{A data frame in which rows define the independent experiments (samples) and columns the features (variables), with the batch in the first column.}
}
\value{
Three datasets, adjusted by either of applied statistical algorithms prepared in \bold{csv} and together with series of plot such as \emph{PCA} plot and \emph{Scree plot} compiled into a \bold{PDF} file are saved in the working directory. \emph{RLA} plots are represented in the \bold{Viewer} panel of \bold{rstudio}.
}
\description{
This function performs batch effect adjustment via three statistical models, namely two-stage procedure as described by M. Giordan 2013 and/or empirical Bayes methods in two setting of parametric and non-parametric as described by Johnson et al. 2007 ( see also "sva") . Meanwhile, the graphical inferences in the context of unsupervised learning algorithms create visual inspection to inform users about the spatial separation of the sample sets analyzed in the different analytical runs alongside the distribution of the features (variables) in the sample sets and across multiple batches.\cr
}
\details{
Zero and NA values are not allowed. optionally missing value can be imputed by \code{emvf} and /or \code{emvd}, functions implemented in the \code{dbnorm} package. Input data must be normalized prior.
}
\examples{
\dontrun{
batch<- rep(gl(5,10,labels = c(1:5)),1)
y<- matrix(rnorm(5000),nrow=50)
f<-data.frame(batch,y)
Visdbnorm(f)}

}
\references{
M. Giordan 2013 \url{https://link.springer.com/article/10.1007/s12561-013-9081-1}\cr
Johnson et al. 2007 \url{http://www.ncbi.nlm.nih.gov/pubmed/16632515}\cr
Leek et al. 2012 \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3307112/}
}
\keyword{Visualization}
\keyword{across}
\keyword{and}
\keyword{batch}
\keyword{normalization}
