% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNorm_commons.R
\name{visDistData}
\alias{visDistData}
\title{Visualising distribution dataset}
\usage{
visDistData(DBdata, type, t, xl, yl)
}
\arguments{
\item{DBdata}{a distribution dataset}

\item{type}{plot by frequence / probability}

\item{t}{title of plot}

\item{xl}{description of x-asis}

\item{yl}{description of y-asis}
}
\description{
Visualising distribution dataset
}
\details{
This function generates distribution data based on predefined distribution. 
The purpose of this function is to enable to normalize arbitrary distributions 
into a standard distribution.
}
\examples{
# visualising a distribution data
DBdata1 <- genDistData(DArray1, 500)
visDistData(DBdata1, "F", "DArray1", "Range", "Frequence")
visDistData(DBdata1, "P", "DArray1", "Range", "Probability")

}
\author{
Qinxue Meng
}

