% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNorm_commons.R
\name{visFitting}
\alias{visFitting}
\title{Visualising fitting results on the input distribution}
\usage{
visFitting(DBdata, t, xl, yl)
}
\arguments{
\item{DBdata}{a distribution dataset}

\item{t}{title of plot}

\item{xl}{description of x-asis}

\item{yl}{description of y-asis}
}
\description{
Visualising fitting results on the input distribution
}
\details{
The function visualizes data distribution and corresponding fitting 
function so as to provide an intutive way to evaluate the 
performance of fitting function.
}
\examples{
# visualising fitting results on DArray1's distribution
visFitting(DBdata1, "DArray1", "Range", "Probability")

}
\author{
Qinxue Meng
}

