% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBNorm_Mapping.R
\name{conNormalizer}
\alias{conNormalizer}
\title{Normalizing a target data array to a basis array based on their distributions}
\usage{
conNormalizer(tg, bs)
}
\arguments{
\item{tg}{a target data array}

\item{bs}{a basis data array}
}
\value{
A normalized target data array with 
the same distribution with the basis data array
}
\description{
Normalizing a target data array to a basis array based on their distributions
}
\details{
The function maps a target data array to a basis array based on their distributions 
and the basis data array can be an arbitary data array or a standard distribution 
such as normal distribution.
}
\examples{
# Normalize DArray1 to DArray3
DArray1 = conNormalizer(DArray1, DArray3)

}
\author{
Qinxue Meng, Paul Kennedy
}

