% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbnorm.R
\name{dbnormPcom}
\alias{dbnormPcom}
\title{drift across batch normalization via Parametric- ComBat model and visualization base on unsupervised learning algorithm and regression analysis}
\usage{
dbnormPcom(m)
}
\arguments{
\item{m}{A data frame in which rows define the independent experiments (samples) and columns the features (variables), with the batch levels in the first column.}
}
\value{
Several graphs compiled into a \bold{PDF} file are a \emph{PCA} score plot, \emph{Scree} plot and a  \emph{correlation} plot for raw and corrected data. Also, the \emph{RLA} plots for each dataset visualized in the \bold{Viewer} panel in the \bold{rstudio} console.\cr
Files saved as \bold{csv} in the working directory are a dataset corrected by the applied model. Also, a two column matrix for Adjusted R-Square raw and corrected dataset and a table summarizing the maximum score.
}
\description{
This function allows you to adjust the data for across batch signal drift or batch effect  parametric Empirical Bayes approach as described by \emph{Johnson et al.(2007)} and in \emph{sva} package as explained by \emph{Leek et al.(2012)}. emph{dbnormPcom}  includes advanced statistical tools to inspect the structure and quality of high throughput experiment both in macroscopic and microscopic scales at the level of sample sets and metabolic feature, respectively. Notably, using this function users applied unsupervised learning algorithm to visualize the most variance explained by the two first components in the different set of samples analyzed in the entire experiment in the raw and corrected data. In parallel, linear association of feature (variable) and batch level has been estimated and visualized by a correlation plot. In fact, estimated \emph{Adjusted- R squared} is considered to define the level of dependency of feature (variable) to the batch level in the raw and corrected datasets. Besides, for quick notification about the performance of the applied model a maximum variability detected in either of datasets is reported as a score. This score notify the consistency of model performance for all detected features (variables).\cr
}
\details{
Zero and NA values are not allowed. Optionally missing value can be imputed by functions implemented in \code{dbnorm} package. Input must be normalized and transformed prior.
}
\examples{

batch<- rep(gl(3,7,labels = c(1:3)),1)
y<- matrix(rnorm(2100),nrow=21)
m<-data.frame(batch,y)
}
\references{
Johnson et al. (2007) \emph{http://www.ncbi.nlm.nih.gov/pubmed/16632515}\cr
Leek et al. (2012) \emph{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3307112/}
}
\keyword{ComBat-Parametric}
\keyword{analysis}
\keyword{and}
\keyword{data}
\keyword{for}
\keyword{normalized}
\keyword{regression}
\keyword{unsupervised}
\keyword{via}
