% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbnorm.R
\name{Visdbnorm}
\alias{Visdbnorm}
\title{Visualization and normalization of signal drift across batches\cr
\emph{Visdbnorm} performs batch effect adjustment via three statistical models implemented in the \code{dbnorm}, namely two-stage procedure as described by \emph{Giordan (2013)} and/or empirical Bayes methods in two setting of parametric and non-parametric as described by \emph{Johnson et al.(2007)} and in \emph{sva} package by \emph{Leek et al.(2012)}. Meanwhile, the graphical inferences in the context of unsupervised learning algorithms create visual inspection to inform users about the spatial separation of the sample sets analyzed in the different analytical runs alongside the distribution of the features (variables) in the raw and treated datasets. This function is suggested for less than 2000 features (variables) to speed up the computational process. \cr}
\usage{
Visdbnorm(f)
}
\arguments{
\item{f}{A data frame in which rows define the independent experiments (samples) and columns the features (variables), with the batch in the first column.}
}
\value{
Three datasets, adjusted by either of applied statistical algorithms prepared in \bold{csv} and together with series of plot such as \emph{PCA} plot and \emph{Scree plot} compiled into a \bold{PDF} file are saved in the working directory. \emph{RLA} plots are represented in the \bold{Viewer} panel of \bold{rstudio}.
}
\description{
Visualization and normalization of signal drift across batches\cr
\emph{Visdbnorm} performs batch effect adjustment via three statistical models implemented in the \code{dbnorm}, namely two-stage procedure as described by \emph{Giordan (2013)} and/or empirical Bayes methods in two setting of parametric and non-parametric as described by \emph{Johnson et al.(2007)} and in \emph{sva} package by \emph{Leek et al.(2012)}. Meanwhile, the graphical inferences in the context of unsupervised learning algorithms create visual inspection to inform users about the spatial separation of the sample sets analyzed in the different analytical runs alongside the distribution of the features (variables) in the raw and treated datasets. This function is suggested for less than 2000 features (variables) to speed up the computational process. \cr
}
\details{
Zero and NA values are not allowed. Optionally missing value can be imputed by \code{emvf} or \code{emvd}, functions implemented in the \code{dbnorm} package. Input data must be normalized prior.
}
\examples{

batch<- rep(gl(3,7,labels = c(1:3)),1)
y<- matrix(rnorm(2100),nrow=21)
f<-data.frame(batch,y)
}
\references{
M.Giordan (2013) \emph{https://link.springer.com/article/10.1007/s12561-013-9081-1}\cr
Johnson et al. (2007) \emph{http://www.ncbi.nlm.nih.gov/pubmed/16632515}\cr
Leek et al. (2012) \emph{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3307112/}
}
\keyword{Visualization}
\keyword{across}
\keyword{and}
\keyword{batch}
\keyword{normalization}
